Add files via upload
Principal Component Analysis (PCA) is a statistical technique used to analyze the variation in a set of variables. It is often used to reduce the dimensionality of large data sets, by transforming the data into a new coordinate system, where the new variables are orthogonal linear combinations of the original variables, and the first few of them capture most of the variation in the data. These new variables, called principal components, are ranked in order of the amount of variation they explain in the data, and the ones with the largest explained variance are selected for further analysis. PCA is widely used in fields such as computer vision, image processing, and bioinformatics.